feat: Use websocket for notifications #195
