+++ frontend/src/layout/navbar/Navbar.tsx
@@ -6,200 +6,16 @@
  * @example <Notification/>
  */
 
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import { useAuth } from "contexts/authContext";
 import { useNavigate } from "react-router-dom";
 
 // components
-import Button from "components/button/Button";
 import Select from "components/select/Select";
-
-import Svg from "assets/svg/Svg";
 import style from "./navbar.module.scss";
-
-import _ from "lodash";
-
-import audioURL from "assets/audio/notification-a.mp3";
-import useApi from "hooks/useApi";
-
-import View from "pages/notifications/popup/View";
-import Autofill from "components/input/Autofill";
 import useOutsideClick from "hooks/useOutsideClick";
 
-const Notification = () => {
-  /**
-   * active state for notification contents
-   */
-  const { currentUser, socket, currentNotificationsRef } = useAuth();
-  const { NotificationApi } = useApi();
-  const navigate = useNavigate();
-
-  const [notificationContentActive, setNotificationContentActive] =
-    useState(false);
-  const [notificationPopupActive, setNotificationPopupAcitve] = useState(false);
-
-  const [isNotificationLoading, setIsNotifiationLoading] = useState(false);
-  const [isNotificationContenLoading, setIsNotifiationContenLoading] =
-    useState(false);
-  const [notification, setNotification] = useState<any>();
-
-  const notificationtRef = useRef<HTMLDivElement>(null);
-
-  const audio = new Audio(audioURL);
-
-  function handleMousedown(e: MouseEvent) {
-    if (
-      notificationtRef.current &&
-      !notificationtRef.current.contains(e.target as Node)
-    ) {
-      setNotificationContentActive(false);
-    }
-  }
-
-  useEffect(() => {
-    if (isNotificationLoading) {
-      NotificationApi.RNotifications({
-        type: "received",
-        user: currentUser._id,
-        checked: false,
-      }).then((res) => {
-        if (res) {
-          // audio.play().catch((e: any) => {
-          //   // console.log(e);
-          // });
-          currentNotificationsRef.current = res;
-          setIsNotifiationLoading(false);
-        }
-      });
-    }
-
-    return () => {};
-  }, [isNotificationLoading]);
-
-  useEffect(() => {
-    const socketHandler = () => setIsNotifiationLoading(true);
-    if (socket) {
-      socket.on("checkNotifications", socketHandler);
-      socket.emit("login", {
-        academyId: currentUser.academyId,
-        userId: currentUser.userId,
-      });
-    }
-
-    return () => {};
-  }, []);
-
-  useEffect(() => {
-    document.addEventListener("mousedown", handleMousedown);
-    return () => {
-      document.removeEventListener("mousedown", handleMousedown);
-    };
-  }, []);
-
-  useEffect(() => {
-    if (isNotificationContenLoading) {
-      setIsNotifiationContenLoading(false);
-    }
-    return () => {};
-  }, [isNotificationContenLoading]);
-
-  const notifications = () => {
-    return currentNotificationsRef.current.map((notification: any) => {
-      return (
-        <div className={style.item} style={{ marginBottom: "12px" }}>
-          <div
-            className={style.description}
-            onClick={() => {
-              setNotification(notification);
-              setNotificationPopupAcitve(true);
-            }}
-          >
-            {notification.category && (
-              <span className={style.type}>[{notification.category}]</span>
-            )}
-            {notification.title}
-          </div>
-          <Button
-            type="ghost"
-            onClick={(e: any) => {
-              NotificationApi.UCheckNotification(notification._id)
-                .then(() => {
-                  currentNotificationsRef.current.splice(
-                    _.findIndex(
-                      currentNotificationsRef.current,
-                      (x: any) => x._id === notification._id
-                    ),
-                    1
-                  );
-                  setIsNotifiationContenLoading(true);
-                })
-                .catch((err) => {
-                  alert(err.response.data.message);
-                });
-            }}
-            style={{
-              border: 0,
-              color: "gray",
-            }}
-          >
-            x
-          </Button>
-        </div>
-      );
-    });
-  };
-
-  return (
-    <>
-      <div className={style.notification} ref={notificationtRef}>
-        <div
-          className={`${style.icon} ${
-            currentNotificationsRef.current.length > 0 && style.active
-          }`}
-          onClick={() => {
-            setNotificationContentActive((prev) => !prev);
-          }}
-          data-count={
-            currentNotificationsRef.current.length > 0
-              ? currentNotificationsRef.current.length
-              : ""
-          }
-        >
-          <Svg type="notification" width="20px" height="20px" />
-        </div>
-
-        {notificationContentActive && !isNotificationLoading && (
-          <>
-            <div className={style.contents}>
-              <div className={style.title}>알림</div>
-              <div className={style.item_box}>
-                {!isNotificationContenLoading && notifications()}
-              </div>
-              <div className={style.button}>
-                <Button
-                  type="ghost"
-                  onClick={(e: any) => {
-                    setNotificationContentActive(false);
-                    navigate("/notifications");
-                  }}
-                >
-                  모두보기
-                </Button>
-              </div>
-            </div>
-            {notificationPopupActive && (
-              <View
-                setState={setNotificationPopupAcitve}
-                data={notification}
-                type={"received"}
-              />
-            )}
-          </>
-        )}
-      </div>
-    </>
-  );
-};
+import Notification from "./Notification";
 
 type Props = { title?: string };
 
@@ -239,25 +55,6 @@ const Navbar = (props: Props) => {
     });
     setUsers(users);
   }, [currentSchool, currentSeason]);
-
-  // useEffect(() => {
-  //   if (!currentSchool || !currentSeason) return;
-
-  //   RegistrationApi.RRegistrations({
-  //     season: currentSeason._id,
-  //     school: currentSchool.school,
-  //   })
-  //     .then((result) => {
-  //       const users = result.map((user: any) => {
-  //         return {
-  //           text: `${user.userName} / ${user.userId}`,
-  //           value: user.userId,
-  //         };
-  //       });
-  //       setUsers(users);
-  //     })
-  //     .catch((error) => console.error(error));
-  // }, [currentSchool, currentSeason]);
 
   return (
     <div className={style.navbar_container}>
